+++ suite_test.go
@@ -81,6 +81,10 @@ func SharedSuite(t *testing.T, client *quark.Client) {
 		testUpdateZeroValues(ctx, t, client)
 	})
 
+	t.Run("JoinOnSecurity", func(t *testing.T) {
+		testJoinOnSecurity(ctx, t, client)
+	})
+
 	t.Run("Events", func(t *testing.T) {
 		testEvents(ctx, t, client)
 	})